Give release tasks separate folder
