+++ tasks/release/release.js
@@ -1,12 +1,12 @@
 'use strict';
 
 var gulp = require('gulp');
-var utils = require('./utils');
+var utils = require('../utils');
 
 var releaseForOs = {
-    osx: require('./release_osx'),
-    linux: require('./release_linux'),
-    windows: require('./release_windows'),
+    osx: require('./osx'),
+    linux: require('./linux'),
+    windows: require('./windows'),
 };
 
 gulp.task('release', ['build'], function () {
